+++ index.php
@@ -4,6 +4,7 @@ require_once('config.php');
 require_once('functions.php');
 require_once('init.php');
 require_once('data.php');
+require_once('getwinner.php');
 
 $major_indexes = [
   'categories' => $categories,
